fix: Solve memory leak and exit status problem
There used to be a memory leak when non-builtin commands were called more than 1 time.
This was because of allocating "dir" to cmd.command in check_path function
but freeing it in the child process, which does not affect the parent process.
So, I freed cmd.command in the parent process when cmd.command is not equal to
s->curr_cmds->curr_tok->token[0]. That is, when "dir" was allocated to cmd.command.

ft_cd, ft_export, ft_unset had exit status issues.
Exit status did not change to 1 when the commands were failed because the command functions
did not return right after changing the exit status to 1. Instead, the functions went on
and the exit status was re-initialized to 0.