fix: Uninitialize s->flag to 0
Initializing flag in t_state struct to 0 in remove_quotes_backslashes function made it impossible to print out a 'single quote error' later in the code because the error print is triggered by checking if the flag in t_state struct is 0 or not. Initialization was merely a mistake of mine.